[FIX] CameraInfo Height Width
Adding Height and Width definitions to the switch case in ZedCameraROS. This populates the CameraInfo messages for the cameras, which are otherwise set to 0. Though width_ and height_ are variables which are instantiated and referenced, they are never assigned a value in the ZedCameraROS class. The values cannot be inherited either as they are private within StereoCamera.


Thank you for writing this package :)